+++ MS_LIBRARY/trunk/src/org/yeastrc/ms/parser/sqtFile/sequest/SequestSQTFileReader.java
@@ -190,7 +190,7 @@ public class SequestSQTFileReader extends SQTFileReader<SequestSearchScan> {
         public String getServerName() {
             return scan.getServerName();
         }
-        public BigDecimal getTotalIntensity() {
+        public Double getTotalIntensity() {
             return scan.getTotalIntensity();
         }
     }
